Add links to empty URL's
URL's that should be redirecting to our Trezor Wiki or Trezor Blog are also searchable by the Google, but the link is empty. See the example: https://trezor.io/support/general/how-to-reset-a-lost-or-forgotten-pin/

Since it is not possible to reach this link directly from Trezor Troubleshooter at https://trezor.io/support/ but only by specifically searching through search engines, it could be enough to add the link to those empty URL's. This hot fix seems to be acceptable also for the reason that the current troubleshooter is being replaced by new one.